Ensuring blank values gets set to nil for untyped attributes.
This has always been a bit of an annoyance for me with ActiveRecord,
if you don't set blank attributes to nil you can't reliably query for
them in a SQL database (is it an empty string?, does it have one space,
many spaces or is it NULL?).

Please tell me if this does not work for you. In my experience I've
never found a situation where I needed to save a blank string in a
database.